@@ -82,48 +90,49 @@
     "react-redux": "^5.0.6",
     "react-router": "^3.0.2",
     "react-router-redux": "^4.0.8",
-    "react-select": "^1.0.0-rc.10",
     "redux": "^3.7.2",
     "redux-saga": "^0.15.6",
     "redux-thunk": "^2.2.0",
+    "react-select": "^1.2.4",
     "sanitize-filename": "^1.6.1",
-    "sanitize-html": "^1.14.1",
-    "scmp": "^1.0.2",
-    "semver": "^5.4.1",
-    "serve-favicon": "^2.4.5",
-    "string-to-stream": "^1.1.0",
     "tinymce": "^4.9.2",
+    "sanitize-html": "^1.20.1",
+    "scmp": "^1.0.2",
+    "semver": "^6.0.0",
+    "serve-favicon": "^2.5.0",
+    "string-to-stream": "^1.1.1",
     "vkey": "^1.0.1",
-    "watchify": "^3.9.0",
-    "xhr": "^2.4.0"
+    "watchify": "^3.11.1",
+    "xhr": "^2.5.0"
   },
   "devDependencies": {
-    "browserify-middleware": "^7.1.0",
-    "codeclimate-test-reporter": "^0.5.0",
-    "connect-mongo": "^1.3.2",
-    "core-assert": "^0.2.1",
-    "cross-env": "^3.1.4",
+    "@babel/register": "^7.4.4",
+    "browserify-middleware": "^8.1.1",
+    "codeclimate-test-reporter": "^0.5.1",
+    "connect-mongo": "^2.0.3",
+    "core-assert": "^1.0.0",
+    "cross-env": "^5.2.0",
     "disc": "^1.3.3",
-    "enzyme": "^2.8.2",
-    "eslint": "^3.17.0",
+    "enzyme": "^2.9.1",
+    "eslint": "^5.16.0",
     "eslint-config-keystone": "^3.0.0",
     "eslint-config-keystone-react": "^1.0.0",
-    "eslint-config-postcss": "^2.0.2",
-    "eslint-plugin-react": "^6.9.0",
+    "eslint-config-postcss": "^3.0.7",
+    "eslint-plugin-react": "^7.13.0",
     "istanbul": "^0.4.5",
-    "keystone-email": "^1.0.5",
+    "keystone-email": "^1.1.0",
     "keystone-nightwatch-e2e": "^0.2.14",
     "mocha": "^3.5.3",
     "must": "^0.13.4",
-    "proxyquire": "^1.8.0",
+    "proxyquire": "^2.1.0",
     "react-addons-test-utils": "^15.6.2",
     "react-engine": "^4.5.1",
-    "rimraf": "^2.6.2",
-    "sinon": "^2.3.2",
-    "superagent": "^3.8.3",
-    "supertest": "^3.0.0",
-    "uglify-js": "^2.8.18",
-    "updtr": "^1.0.0",
+    "rimraf": "^2.6.3",
+    "sinon": "^2.4.1",
+    "superagent": "^5.0.5",
+    "supertest": "^3.4.2",
+    "uglify-js": "^3.5.13",
+    "updtr": "^3.1.0",
     "watch": "^1.0.2"
   },
   "browserify": {
