@@ -84,7 +84,7 @@ var EditForm = React.createClass({
 				props.isValid = false;
 			}
 		}
-		props.value = this.state.values[field.path] || field.defaultValue;
+		props.value = this.state.values[field.path] === undefined ? field.defaultValue : this.state.values[field.path];
 		props.values = this.state.values;
 		props.onChange = this.handleChange;
 		props.mode = 'edit';
