@@ -39,5 +39,5 @@ exports.spacing = {
 
 // table constants
 
-exports.TABLE_CONTROL_COLUMN_WIDTH = 26;  // icon + padding
+exports.TABLE_CONTROL_COLUMN_WIDTH = 26; // icon + padding
 exports.NETWORK_ERROR_RETRY_DELAY = 500; // in ms
