@@ -37,8 +37,8 @@ boolean.prototype.validateInput = function (data, callback) {
 boolean.prototype.validateRequiredInput = function (item, data, callback) {
 	var value = this.getValueFromData(data);
 	var result = value && value !== 'false'
-			? true
-			: false;
+		? true
+		: false;
 	utils.defer(callback, result);
 };
 
