@@ -86,10 +86,9 @@ module.exports = function getRelated (paths, callback, nocollapse) {
 							item.populateRelated(options.related, done);
 						};
 					}),
-						function (err) {
-							done(err, results);
-						}
-					);
+					function (err) {
+						done(err, results);
+					});
 				});
 			} else {
 				query.exec(done);
