@@ -23,12 +23,12 @@ var ADAPTER_COMPATIBILITY_LEVEL = 1;
 	Storage Adapters can specify additional schema properties.
 */
 var SCHEMA_TYPES = {
-//	filename: String,       // the filename, without the full path
-	size: Number,           // the size of the file
-	mimetype: String,       // the mime type of the file
-	path: String,           // the path (e.g directory) the file is stored in; not the full path to the file
-	originalname: String,   // the original (uploaded) name of the file; useful when filename generated
-	url: String,            // publicly accessible URL of the stored file
+//	filename: String, // the filename, without the full path
+	size: Number, // the size of the file
+	mimetype: String, // the mime type of the file
+	path: String, // the path (e.g directory) the file is stored in; not the full path to the file
+	originalname: String, // the original (uploaded) name of the file; useful when filename generated
+	url: String, // publicly accessible URL of the stored file
 };
 
 /**
